Fix postgres / redis ports in production
It's great that we attempted to run migrations and failed early, instead of having a full-blown downtime.

The reason the ports are non-standard on the local dev side is to allow running multiple projects at the same time, without the port conflicts. And in production, the ports are always standard