Fix ELO graphs
Update the moment.js import to explicitly use the pre-bulilt minimised version (instead of the one auto-generated by the CDN - the auto-generated version does not have a stable SRI hash). Also update the integrity hash to the official hash for this version of this file.

I ran the site locally, and with this patch in place I can once again see the ELO graphs.