Optimize the bot's recent results query
This makes sure it uses an index instead of scanning a few millions of rows

Also, this makes a breaking change in the sorting. We used to sort by the date the match actually started, and now we're effectively sorting by the date the request was created (because we're sorting by id). That's the only way I was able to make the index work, but I think it's a decent compromise, considering it's still possible to change the sort manually in the table if you really want the old behavior